@@ -1,6 +1,9 @@
 {
   "private": true,
   "scripts": {
+    "build-force": "vite build --force",
+    "build": "vite build",
+    "preview": "vite preview",
     "dev": "npm run development",
     "development": "mix",
     "watch": "mix watch",
@@ -11,26 +14,31 @@
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.15.1",
-    "axios": "^0.21.1",
-    "bootbox": "^5.5.2",
-    "bootstrap": "^4.5.3",
+    "@vitejs/plugin-vue": "^4.1.0",
+    "@vue/compat": "^3.1.0",
+    "axios": "^1.3.6",
+    "bootstrap": "^5.2.3",
+    "bootstrap-icons": "^1.10.4",
     "bootstrap-select": "^1.13.18",
     "cross-env": "^7.0.3",
-    "gameap-file-manager": "^2.0.14",
+    "gameap-file-manager": "^3.0.0",
     "jquery": "^3.5.1",
     "laravel-mix": "^6.0.0",
     "lodash": "^4.17.20",
     "moment": "^2.29.1",
     "popper.js": "^1.15.0",
     "sass": "^1.32.0",
-    "sass-loader": "^8.0.0",
-    "vue": "^2.6.12",
-    "vue-select": "^2.5.1",
-    "vue2-datepicker": "^3.8.2"
+    "sass-loader": "^9.0.0",
+    "vue": "^3.1.0",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
+    "@originjs/vite-plugin-commonjs": "^1.0.3",
+    "@vue/compiler-sfc": "^3.1.0",
+    "laravel-vite-plugin": "^0.7.4",
+    "naive-ui": "^2.34.3",
     "resolve-url-loader": "^4.0.0",
-    "vue-loader": "^15.9.8",
-    "vue-template-compiler": "^2.6.14"
+    "vite": "^4.3.1",
+    "vue-loader": "^16.0.0"
   }
 }
