@@ -1,5 +1,5 @@
-
-window._ = require('lodash');
+import _ from 'lodash'
+window._ = _
 
 /**
  * We'll load jQuery and the Bootstrap jQuery plugin which provides support
@@ -7,13 +7,11 @@ window._ = require('lodash');
  * code may be modified to fit the specific needs of your application.
  */
 
-try {
-    window.$ = window.jQuery = require('jquery');
+window.bootstrap = {};
 
-    require('bootstrap');
-} catch (e) {}
-
-require('./i18n/i18n');
+import {Modal, Tooltip} from 'bootstrap';
+window.bootstrap.Modal = Modal;
+window.bootstrap.Tooltip = Tooltip;
 
 /**
  * We'll load the axios HTTP library which allows us to easily issue requests
@@ -21,9 +19,10 @@ require('./i18n/i18n');
  * CSRF token as a header based on the value of the "XSRF" token cookie.
  */
 
-window.axios = require('axios');
+import axios from 'axios'
+window.axios = axios
 
-window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
+window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest'
 
 /**
  * Next we will register the CSRF Token as a common header with Axios so that
@@ -39,9 +38,6 @@ if (token) {
     console.error('CSRF token not found: https://laravel.com/docs/csrf#csrf-x-csrf-token');
 }
 
-window.bootbox = require('bootbox');
-require('bootstrap-select');
+import 'bootstrap-select';
 
-import Vue from 'vue';
-
-window.Vue = Vue;
+import 'popper.js'
