@@ -33,7 +33,7 @@ return [
      * 2 - one file manager window with directories tree module
      * 3 - two file manager windows
      */
-    'windowsConfig' => 2,
+    'windowsConfig' => 1,
 
     /**
      * Middleware
