@@ -1,13 +1,13 @@
 <template>
     <div>
-        <div id="dedicatedServerForm" class="form-group">
+        <div id="dedicatedServerForm" class="mb-3">
             <label for="ds_id" class="control-label">{{ trans('labels.ds_id') }}</label>
             <select id="ds_id" v-bind:name="dsIdFieldName" class="form-control" v-model="selectedDs">
                 <option v-for="(dsName, dsId) in dsList" v-bind:value="dsId">{{ dsName }}</option>
             </select>
         </div>
 
-        <div class="form-group">
+        <div class="mb-3">
             <label for="server_ip" class="control-label">{{ trans('labels.server_ip') }}</label>
             <select id="server_ip" v-bind:name="serverIpFieldName" class="form-control" v-model="selectedIp">
                 <option v-for="ip in ipList" v-bind:value="ip">{{ ip }}</option>
