@@ -1,13 +1,13 @@
 <template>
     <div>
-        <div class="form-group">
+        <div class="mb-3">
             <label for="game_id" class="control-label">{{ trans('labels.game_id') }}</label>
             <select id="game_id" name="game_id" class="form-control" v-model="selectedGameCode">
                 <option v-for="(gameName, gameCode) in games" v-bind:value="gameCode">{{ gameName }}</option>
             </select>
         </div>
 
-        <div class="form-group">
+        <div class="mb-3">
             <label for="game_mod_id" class="control-label">{{ trans('labels.game_mod_id') }}</label>
             <select id="game_mod_id" name="game_mod_id" class="form-control" v-model="selectedMod">
                 <option v-for="gameMod in gameModsList" v-bind:value="gameMod.id">{{ gameMod.name }}</option>
