@@ -1,20 +1,15 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
-
+import { createStore } from 'vuex';
 import dedicatedServers from './dedicatedServers';
 import servers from './servers';
 import games from './games';
 import gameMods from './gameMods';
-
 import activeTab from './activeTab';
 
 // Rcon
-import rconConsole from './rcon/console'
-import rconPlayers from './rcon/players'
+import rconConsole from './rcon/console';
+import rconPlayers from './rcon/players';
 
-Vue.use(Vuex);
-
-export default new Vuex.Store({
+const store = createStore({
     modules: {
         dedicatedServers,
         servers,
@@ -27,4 +22,6 @@ export default new Vuex.Store({
         rconConsole,
         rconPlayers,
     },
-})
+});
+
+export default store;
