@@ -61,7 +61,7 @@
         </div>
 
         <div class="col-md-12 mt-4">
-            <div class="form-group">
+            <div class="mb-3">
                 {{ Form::submit(__('main.create'), ['class' => 'btn btn-success']) }}
             </div>
         </div>
