@@ -21,20 +21,18 @@
         {{ Form::bsPassword('password_confirmation') }}
         {{ Form::bsText('name') }}
 
-        <div class='form-group'>
+        <div class='mb-3'>
             {{ Form::label('roles', __('users.roles'), ['class' => 'control-label']) }}
 
-            {{ Form::select(
-                'roles[]',
-                $roles->pluck('title', 'name'),
-                null,
-                ['id' => 'roles', 'multiple' => 'multiple', 'class' => 'form-control selectpicker']
-            ) }}
+            <gameap-select
+                    name="roles[]"
+                    :options="{{ json_encode($roleOptions) }}">
+            </gameap-select>
         </div>
     </div>
 
     <div class="col-md-12">
-        <div class="form-group">
+        <div class="mb-3">
             {{ Form::submit(__('main.create'), ['class' => 'btn btn-success']) }}
         </div>
     </div>
