@@ -8,11 +8,11 @@
                 <form class="form-horizontal" method="POST" action="{{ route('login') }}">
                     {{ csrf_field() }}
 
-                    <div class="card-header">{{ __('auth.sign_in') }}</div>
+                    <div class="card-header p-4">{{ __('auth.sign_in') }}</div>
 
-                    <div class="card-body">
+                    <div class="card-body p-4">
 
-                        <div class="form-group{{ $errors->has('login') ? ' has-error' : '' }}">
+                        <div class="mb-3{{ $errors->has('login') ? ' has-error' : '' }}">
                             <label for="login" class="col-md-12">{{ __('auth.username_email') }}</label>
 
                             <div class="col-md-12">
@@ -26,7 +26,7 @@
                             </div>
                         </div>
 
-                        <div class="form-group{{ $errors->has('password') ? ' has-error' : '' }}">
+                        <div class="mb-3{{ $errors->has('password') ? ' has-error' : '' }}">
                             <label for="password" class="col-md-12">{{ __('auth.password') }}</label>
 
                             <div class="col-md-12">
@@ -40,7 +40,7 @@
                             </div>
                         </div>
 
-                        <div class="form-group">
+                        <div class="mb-3">
                             <div class="col-md-6 col-md-offset-4">
                                 <div class="checkbox">
                                     <label>
@@ -51,8 +51,8 @@
                         </div>
                     </div>
 
-                    <div class="card-footer text-muted">
-                        <div class="form-group">
+                    <div class="card-footer p-4 text-muted">
+                        <div class="mb-3">
                             <div class="col-md-12">
                                 <div class="row">
                                     <div class="col-md-12">
