@@ -7,18 +7,18 @@
             <div class="card mb-3 mt-3">
                 <form class="form-horizontal" method="POST" action="{{ route('password.request') }}">
                     <div class="card-header">{{ __('auth.reset_password') }}</div>
-    
+
                     <div class="card-body">
                             {{ csrf_field() }}
-    
+
                             <input type="hidden" name="token" value="{{ $token }}">
-    
-                            <div class="form-group{{ $errors->has('email') ? ' has-error' : '' }}">
+
+                            <div class="mb-3{{ $errors->has('email') ? ' has-error' : '' }}">
                                 <label for="email" class="col-md-4 control-label">{{ __('auth.email') }}</label>
-    
+
                                 <div class="col-md-6">
                                     <input id="email" type="email" class="form-control" name="email" value="{{ $email or old('email') }}" required autofocus>
-    
+
                                     @if ($errors->has('email'))
                                         <span class="help-block">
                                             <strong>{{ $errors->first('email') }}</strong>
@@ -26,13 +26,13 @@
                                     @endif
                                 </div>
                             </div>
-    
-                            <div class="form-group{{ $errors->has('password') ? ' has-error' : '' }}">
+
+                            <div class="mb-3{{ $errors->has('password') ? ' has-error' : '' }}">
                                 <label for="password" class="col-md-4 control-label">{{ __('auth.password') }}</label>
-    
+
                                 <div class="col-md-6">
                                     <input id="password" type="password" class="form-control" name="password" required>
-    
+
                                     @if ($errors->has('password'))
                                         <span class="help-block">
                                             <strong>{{ $errors->first('password') }}</strong>
@@ -40,12 +40,12 @@
                                     @endif
                                 </div>
                             </div>
-    
-                            <div class="form-group{{ $errors->has('password_confirmation') ? ' has-error' : '' }}">
+
+                            <div class="mb-3{{ $errors->has('password_confirmation') ? ' has-error' : '' }}">
                                 <label for="password-confirm" class="col-md-4 control-label">{{ __('auth.confirm_password') }}</label>
                                 <div class="col-md-6">
                                     <input id="password-confirm" type="password" class="form-control" name="password_confirmation" required>
-    
+
                                     @if ($errors->has('password_confirmation'))
                                         <span class="help-block">
                                             <strong>{{ $errors->first('password_confirmation') }}</strong>
@@ -54,9 +54,9 @@
                                 </div>
                             </div>
                     </div>
-    
+
                     <div class="card-footer text-muted">
-                        <div class="form-group">
+                        <div class="mb-3">
                             <div class="col-md-12">
                                 <button type="submit" class="btn btn-light">{{ __('auth.reset_password') }}</button>
                             </div>
