@@ -12,7 +12,7 @@
                         <div class="card-header">{{ __('auth.sign_up') }}</div>
 
                         <div class="card-body">
-                            <div class="form-group{{ $errors->has('login') ? ' has-error' : '' }}">
+                            <div class="mb-3{{ $errors->has('login') ? ' has-error' : '' }}">
                                 <label for="login" class="col-md-12">{{ __('auth.login') }}</label>
 
                                 <div class="col-md-12">
@@ -26,7 +26,7 @@
                                 </div>
                             </div>
 
-                            <div class="form-group{{ $errors->has('email') ? ' has-error' : '' }}">
+                            <div class="mb-3{{ $errors->has('email') ? ' has-error' : '' }}">
                                 <label for="email" class="col-md-12">{{ __('auth.email') }}</label>
 
                                 <div class="col-md-12">
@@ -40,7 +40,7 @@
                                 </div>
                             </div>
 
-                            <div class="form-group{{ $errors->has('password') ? ' has-error' : '' }}">
+                            <div class="mb-3{{ $errors->has('password') ? ' has-error' : '' }}">
                                 <label for="password" class="col-md-12">{{ __('auth.password') }}</label>
 
                                 <div class="col-md-12">
@@ -54,7 +54,7 @@
                                 </div>
                             </div>
 
-                            <div class="form-group">
+                            <div class="mb-3">
                                 <label for="password-confirm" class="col-md-12">{{ __('auth.confirm_password') }}</label>
 
                                 <div class="col-md-12">
@@ -70,7 +70,7 @@
                         </div>
 
                         <div class="card-footer">
-                            <div class="form-group">
+                            <div class="mb-3">
                                 <div class="col-md-12">
                                     <button type="submit" class="btn btn-light">{{ __('auth.sign_up') }}</button>
                                 </div>
