@@ -56,7 +56,7 @@
                 <div class="row">
                     <div class="col-12">
                         <div class="col-12">
-                            <a class="btn btn-primary" data-toggle="collapse" href="#changePassword" role="button" aria-expanded="false" aria-controls="changePassword">
+                            <a class="btn btn-primary" data-bs-toggle="collapse" href="#changePassword" role="button" aria-expanded="false" aria-controls="changePassword">
                                 {{ __('profile.change_password') }}
                             </a>
                         </div>
