@@ -104,7 +104,7 @@
 
                         <div class="row mt-2">
                             <div class="col-md-12">
-                                <div class="form-group">
+                                <div class="mb-3">
                                     {{ Form::submit(__('main.send'), ['class' => 'btn btn-success']) }}
                                 </div>
                             </div>
