@@ -1,5 +1,15 @@
 <div class="row mt-2">
     <div class="col-md-12">
-        <file-manager v-if="activeTab === 'filemanager'" server-id="{{ $server->id }}"></file-manager>
+        <div class="card p-2">
+            <file-manager
+                    v-if="activeTab === 'filemanager'"
+                    :settings="{{ json_encode([
+                    'baseUrl' => url('file-manager/'.$server->id),
+                    'headers' => [
+                        'X-Requested-With' => 'XMLHttpRequest',
+                    ],
+                ]) }}"
+            />
+        </div>
     </div>
 </div>
