@@ -39,7 +39,7 @@
                 @endif
 
                 <div class="mt-2 col-md-12">
-                    <div class="form-group">
+                    <div class="mb-3">
                         {{ Form::submit(__('main.save'), ['class' => 'btn btn-success btn-ico btn-ico-save']) }}
                     </div>
                 </div>
